Product Admin - Securing the views.

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -73,8 +74,13 @@ def product_detail(request, product_id):
     return render(request, "products/product_detail.html", context)
 
 
+@login_required
 def add_product(request):
     """add a product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that.")
+        return redirect(reverse("home"))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -82,10 +88,12 @@ def add_product(request):
             messages.success(request, "Successfully added product!")
             return redirect(reverse("product_detail", args=[product.id]))
         else:
-            messages.error(request, "Failed to add product. Please ensure the form is valid.")
+            messages.error(
+                request, "Failed to add product. \
+                    Please ensure the form is valid.")
     else:
         form = ProductForm()
-    
+
     template = "products/add_product.html/"
     context = {
         "form": form,
@@ -93,8 +101,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """Edit a product in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that.")
+        return redirect(reverse("home"))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == "POST":
@@ -104,7 +117,9 @@ def edit_product(request, product_id):
             messages.success(request, "Successfully updated product!")
             return redirect(reverse("product_detail", args=[product_id]))
         else:
-            messages.error(request, "Failed to update product. Please ensure the form is valid.")
+            messages.error(
+                request, "Failed to update product. \
+                    Please ensure the form is valid.")
     else:
         form = ProductForm(instance=product)
         messages.info(request, f"You are editing {product.name}")
@@ -118,8 +133,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """delete a product from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only store owners can do that.")
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, "Product deleted!")
